Temporarily disable "buf breaking" (#3)
This is needed in order to push the recent change to the Go package,
which `buf breaking` considers to be a breaking change. The Go proto
plugin refuses to generate code that has no Go package unless the user
provides a `-M` option. So all users generating Go code for these files
are already using `-M` or are using something like Buf's "managed mode",
which means changing this option from empty to non-empty won't actually
break anyone in practice.

After the changes are pushed to the BSR, we will undo this change and
re-enable the check.